All experiment repository CRUD operations moved to the DAO
Now all experiment repository CRUD operations are the responsibility of the DAO. The cooked analyzer now only receives valid Experiment Repository objects from the DAO and never instantiates a repository on its own.
This ensures proper initialization of the repository by the DAO rather than creating the object improperly and stuff. Also fixed the bug were a results archive was given to an empty repository handle.